refactor to allow ABC for MLDataset
